pyproject.toml
[project]
name = "autopilot"
version = "5.0.0"
description = "Edge AI multi-agent pipeline with SSE streaming, OTel tracing, and self-healing synthesis"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx[http2]>=0.27.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.100.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=24.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "tenacity>=8.2.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.25.0",
    "prometheus-client>=0.20.0",
    "sse-starlette>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "redis>=5.0.0",
    "google-cloud-aiplatform>=1.40.0",
    "chromadb>=0.4.0",
    "ddgs>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "respx>=0.21.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["workflows/bank_to_ynab/tests"]
pythonpath = ["."]
addopts = "-m 'not integration'"
markers = [
    "integration: tests that call real external APIs (Google Gemini, YNAB)",
]
filterwarnings = [
    "ignore::DeprecationWarning:google.genai",
    "ignore::ResourceWarning",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["autopilot"]
